got rid of extra mongodb library and generalized buildRequest util function
